moving ResolveGenerics to a separate file
